+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 - **Added**: Tuples of up to 16 elements implement Render
 - **Added**: Trailing commas for children and attributes
+- **Changed**: Void elements now won't compile with children
+- **Changed**: Element macro now defaults to non-void
 
 ## [0.1.1]
 - **Changed**: Replaced element list (maybe some got removed, or some got added)